[LocalTest]:[ThreeSum]:[c++]

+++ ThreeSum/ThreeSum.cpp
@@ -15,7 +15,7 @@ public:
         return ans;
 
     sort(nums.begin(), nums.end());
-    for(int i = 0; i < nums.size()-2; i++)
+    for(int i = 0; i < (int)nums.size()-2; i++)
     {
         if (i != 0 && nums[i] == nums[i-1]) continue;
         int l = i+1, r = nums.size()-1;
